backend/requirements.txt
# The Core Engine
pathway[xpack-llm]

# Data Connectors
telethon        # For Telegram
praw            # For Reddit
requests        # For GNews/RSS
beautifulsoup4  # If you need to clean up HTML text
feedparser

# Frontend
streamlit       # For the Dashboard
watchdog        # Helps Streamlit reload faster

# Environment
python-dotenv   # To manage API Keys securely
openai          # Standard OpenAI library
cryptg